fix: smallテストの修正。dummyParamsからテストインスタンスを生成するようにする
