--- src/domain/core/Occupation/Occupation.dummy.ts
@@ -1,16 +1,0 @@
-import { Occupation } from "./Occupation";
-
-/**
- * Occupationのダミーデータ
- */
-export const occupationDummy = Occupation.create({
-  id: 1,
-  name: "エンジニア",
-});
-
-export const occupationDummy2 = Occupation.create({
-  id: 2,
-  name: "デザイナー",
-});
-
-
