+++ src/domain/core/Occupation/repository/OccupationRepository.ts
@@ -1,5 +1,5 @@
 import { Result } from "@/util/result";
-import { Occupation } from "./Occupation";
+import { Occupation } from "../Occupation";
 
 export type FindAllOccupationResult = Result<Occupation[], never>;
 export type FindOccupationResult = Result<Occupation | undefined, never>;
