+++ src/domain/core/User/repository/UserRepository.ts
@@ -1,5 +1,5 @@
 import { Result } from "@/util/result";
-import { User } from "./User";
+import { User } from "../User";
 
 export type FindUserResult = Result<User | undefined, never>;
 export type UpdateUserResult = Result<void, never>;
