--- src/domain/core/WorkLocation/WorkLocation.Dummy.ts
@@ -1,14 +1,0 @@
-import { WorkLocation } from "./WorkLocation";
-
-/**
- * WorkLocationのダミーデータ
- */
-export const workLocationDummy = WorkLocation.create({
-  id: 1,
-  name: "北海道",
-});
-
-export const workLocationDummy2 = WorkLocation.create({
-  id: 2,
-  name: "青森県",
-});
